Refactored code so that peaks are represented as floats and the cluster charge is an int as the most common charge All unit tests pass - some still marked as Known to fail 2) Almost all warnings eliminated with the levels in slewis' intellij

+++ trunk/src/main/java/uk/ac/ebi/pride/spectracluster/cluster/SpectralCluster.java
@@ -36,7 +36,7 @@ public class SpectralCluster implements ISpectralCluster, Equivalent<ISpectralCl
         return id;
     }
 
-    public double getPrecursorMz() {
+    public float getPrecursorMz() {
         guaranteeClean();
         ISpectrum consensusSpectrum1 = getConsensusSpectrum();
         if (consensusSpectrum1 == null)
@@ -44,7 +44,7 @@ public class SpectralCluster implements ISpectralCluster, Equivalent<ISpectralCl
         return consensusSpectrum1.getPrecursorMz();
     }
 
-    public double getPrecursorCharge() {
+    public int getPrecursorCharge() {
         guaranteeClean();
         return getConsensusSpectrum().getPrecursorCharge();
     }
